fix Sonar issues
