+++ src/main/java/org/folio/search/service/reindex/jdbc/MergeInstanceRepository.java
@@ -40,7 +40,7 @@ public class MergeInstanceRepository extends MergeRangeRepository {
 
   private static final String SELECT_BY_UPDATED_QUERY = """
     SELECT id, tenant_id, shared, is_bound_with, json, last_updated_date, is_deleted
-    FROM %s.instance
+    FROM %1$s.instance
     WHERE last_updated_date > ?
     ORDER BY last_updated_date ASC
     """;
